[mv] rename article api controllers
